cart development

+++ src/components/SimpleAccordion.js
@@ -2,7 +2,7 @@ import { useEffect } from 'react'
 import { connect } from 'react-redux'
 import { useSession } from 'next-auth/react'
 
-function SimpleAccordion ({ state, setShowModal, insertProductsOnStore, deleteProduct }) {
+function SimpleAccordion ({ state, setProduct, setModify, modify, insertProductsOnStore, deleteProduct, setSingleProductSelected }) {
   const { data: session } = useSession()
 
   const deleteProductFromRedux = (productToDelete) => {
@@ -49,21 +49,22 @@ function SimpleAccordion ({ state, setShowModal, insertProductsOnStore, deletePr
   const deleteProductApi = async (product) => {
     try {
       const response = await fetch(
-        '/api/products'
+        `/api/products?productId=${product.id}`
         ,
         {
           method: 'DELETE',
-          'Content-Type': 'application/json',
-          Authorization: session.accessToken,
-          body: JSON.stringify({
-            productTitle: product.title
-          })
+          headers: {
+            'Content-Type': 'application/json',
+            Authorization: session.accessToken
+          }
         }
       )
 
       const final = await response.json()
 
-      deleteProduct(deleteProductFromRedux(product))
+      if (final === 'Product Deleted') {
+        deleteProduct(deleteProductFromRedux(product))
+      }
 
       return final
     } catch (err) {
@@ -86,43 +87,54 @@ function SimpleAccordion ({ state, setShowModal, insertProductsOnStore, deletePr
 */
   // ____________API________end_____
 
+  const setDataOnClickModifyButton = (singleProduct) => {
+    setModify(true)
+  console.log(singleProduct.ingredients,'9999')
+  const parsedIngredient = JSON.parse(singleProduct.ingredients)
+    setProduct({
+      title: singleProduct.title,
+      description: singleProduct.description,
+      category: singleProduct.category,
+      price: singleProduct.price,
+      img: singleProduct.img,
+      id: singleProduct.id,
+      restaurantId: state.adminData.restaurantId,
+      ingredients: parsedIngredient 
+
+    })
+  }
+
   const splitAllKeyValueStore = Object.keys(state.productsData).map(key => ({ [key]: state.productsData[key] }))
 
   return (
-  <>
-   <div>
-    {
-     splitAllKeyValueStore.map((category) => {
-       return (
-       <>
-        <div>
-         <h2>{Object.keys(category)[0]}</h2>
-         {Object.values(category).map((arrCategory) => {
-           return (
-           <>
-            <div>
-             {arrCategory.map((singleProduct) => {
-               return (
-               <>
-                <p key={Math.random()}> {singleProduct.title} <button onClick={() => setShowModal(true)} className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded">Modify</button> <button onClick={() => deleteProductApi(singleProduct)} className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded">Delete</button></p>
-               </>
-               )
-             })}
+    <>
+      {
+        splitAllKeyValueStore.map((category) => {
+          return (
+            <div key={Object.keys(category)[0]}>
+            <h2>{Object.keys(category)[0]}</h2>
+            {Object.values(category).map((arrCategory) => {
+              return (
+                  <div key={Math.random()}>
+                    {arrCategory.map((singleProduct) => {
+                      return (
+                        <div key={singleProduct.id} >
 
+                        <p >{singleProduct.title}</p>
+                          <button onClick={ () => { setDataOnClickModifyButton(singleProduct); setSingleProductSelected(singleProduct) }} className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded">Modify</button>
+                          <button onClick={() => deleteProductApi(singleProduct)} className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded">Delete</button>
+                        </div>
+                      )
+                    })}
+                  </div>
+              )
+            })}
             </div>
-           </>
-           )
-         })
-         }
-        </div>
-       </>
 
-       )
-     })
-
-    }
-   </div>
-  </>
+          )
+        })
+      }
+    </>
   )
 }
 
@@ -163,3 +175,42 @@ const mapDispatchToProps = {
 }
 
 export default connect(mapStateToProps, mapDispatchToProps)(SimpleAccordion)
+
+/*
+const setDataOnClickModifyButton = (singleProduct) => {
+
+    function dataURLtoFile (dataurl, filename) {
+      const arr = dataurl.split(','); const mime = arr[0].match(/:(.*?);/)[1]
+      const bstr = atob(arr[1]); let n = bstr.length; const u8arr = new Uint8Array(n)
+      while (n--) {
+        u8arr[n] = bstr.charCodeAt(n)
+      }
+      return new File([u8arr], filename, { type: mime })
+    }
+
+    // Usage example:
+
+    if (singleProduct.img !== '') {
+      const file = dataURLtoFile(singleProduct.img, `${singleProduct.title}.png`)
+      setProduct({
+        title: singleProduct.title,
+        description: singleProduct.description,
+        category: singleProduct.category,
+        price: singleProduct.price,
+        img: file,
+        id: singleProduct.id
+
+      })
+
+    }else{
+
+      setProduct({
+        title: singleProduct.title,
+        description: singleProduct.description,
+        category: singleProduct.category,
+        price: singleProduct.price,
+        id: singleProduct.id
+
+      })
+    }
+*/
